fix two, break one
fixed the
 - vblank failing in normal mode
 - crashing on exit

clearly the drm.irq_enabled=true was responsible for former.
(and maybe drm_vblank_init for the latter)

but also in virtual_hw mode now, all of kms_cursor_crc fails
and in normal mode, majority of it fails(incliding opaque and friends)
plus writeback-check-output fails too for both modes.

Signed-off-by: Sumera Priyadarsini <sylphrenadin@gmail.com>